services/reis/pyproject.toml
[tool.poetry]
name = "REI-S"
version = "0.1.0"
description = ""
authors = ["Patrick Reichmann <patrick.reichmann@codecentric.de>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "<4.0,>=3.12.0"
pydantic = "2.11.7"
pydantic-settings = "^2.7.1"
langchain-core = "0.3.65"
langchain-community = "0.3.25"
langchain-openai = "0.3.24"
langchain-postgres = "0.0.13"
psycopg = {extras = ["binary"], version = "^3.2.9"}
fastapi = {extras = ["standard"], version = "0.115.13"}
azure-search-documents = "11.5.2"
azure-identity = "1.23.0"
unstructured = {extras = ["docx", "md", "pptx", "xlsx"], version = "0.17.2"}
pdfminer-six = "^20250506"
ffmpeg-python = "^0.2.0"
aiofiles = "^24.1.0"
prometheus-fastapi-instrumentator = "^7.0.0"
fastmcp = "^2.8.1"
langchain-ollama = "^0.3.3"

[tool.poetry.group.dev.dependencies]
ruff = "0.12.0"
mypy = "1.16.1"
pre-commit = "4.2.0"
pytest = "8.4.1"
pytest-cov = "6.2.1"
pytest-mock = "3.14.1"
pytest-responses = "0.5.1"
pytest-github-report = "0.0.1"
pytest-error-for-skips = "2.0.2"
faker = "37.4.0"
responses = "0.25.7"

[tool.poetry.scripts]
download-tokenizers = "rei_s.prepare:download"
generate-api-spec = "rei_s.generate_open_api:generate"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    "def __repr__",
    ]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "N", "A", "B"]

[tool.mypy]
strict = true
install_types = true
non_interactive = true
ignore_missing_imports = true
exclude = [
    ".cache",
    ]
